+++ src/optimization/constraint_builder/src/launchgui.py
@@ -5,25 +5,23 @@ This file launches the program
 '''
 
 import tkinter as tk
-import constraint_builder_project_overview
-import constraint_builder_obj_import
-import constraint_builder_opening_screen
+# from screens import *
+import screens.mainmenu
 from enum import Enum, unique, auto
 from typing import List, Dict
-from attrs import define, frozen
-import varname_dataclasses as models
-import constraint_processer as proc
+import processor.models as models
 
 
 
 def main():
+	print("Hi")
 	# Very important to instantiate the object so it can be passed arround
 	projectState = models.ProjectState.createEmptyprojectState()
 
 	root = tk.Tk()
 	root.minsize(width=400, height=400)
 
-	constraint_builder_opening_screen.buildOpeningScreen(root, projectState)
+	screens.mainmenu.buildOpeningScreen(root, projectState)
 
 	root.mainloop()
