+++ src/optimization/constraint_builder/src/screens/standardconstraint.py
@@ -1,8 +1,8 @@
 
 import tkinter as tk
-import constraint_builder_project_overview
-import varname_dataclasses as models
-import constraint_processer as proc
+import screens.projectoverview
+import processor.models as models
+import processor.constraintprocesser as proc
 import copy
 from enum import Enum, unique, auto
 from typing import List, Union, Dict
@@ -177,7 +177,7 @@ def transitionToOverview() -> None:
 		child.destroy()
 
 	# Transition
-	constraint_builder_project_overview.buildProjectOverviewGUI(_passedRoot, _passedProjectState)
+	screens.projectoverview.buildProjectOverviewGUI(_passedRoot, _passedProjectState)
